src-tf/requirements.txt
# tensorflow>=2.13.0  # Generic TensorFlow (use tensorflow-macos for Apple Silicon)
tensorflow-macos>=2.16.0  # TensorFlow for Apple Silicon Macs (M1/M2/M3)
tensorflow-metal>=1.2.0   # Metal Performance Shaders for GPU acceleration on Apple Silicon
tensorflow-addons>=0.21.0
opencv-python>=4.8.0
albumentations>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
tensorboard>=2.13.0 